Test change Accounts

+++ web-api/src/main/java/com/luckygroup/webapi/models/Accounts.java
@@ -10,16 +10,16 @@ public class Accounts {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int id;
+  private Long id;
 
   private String username, email, password;
   private int role;
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
